Move Metadata into its own file like other definition components, and give it a custom to_s method
