+++ lib/holidays/definition/region_definition.rb
@@ -1,26 +1,9 @@
 require "holidays/definition/holiday_rule.rb"
 require "holidays/definition/custom_method.rb"
 require "holidays/definition/test.rb"
+require "holidays/definition/metadata.rb"
 
 module Holidays
-  class Metadata
-    attr_reader :name
-    attr_reader :region
-    
-    def initialize(name, region)
-      @name = name
-      @region = region
-    end
-
-    def self.from_yaml(metadata_def)
-      raise ArgumentError, "Expected region to be defined in metadata block" unless metadata_def["region"]
-
-      Metadata.new(
-        metadata_def['name'],
-        metadata_def['region'].to_sym
-      )
-    end
-  end
   
   # An instance of this class would represent the result of parsing a holiday definition file. Each holiday definition file contains holiday rules, tests, metadata, 
   # and custom methods for a single region.
